Additional checks

+++ backend/service/userService.js
@@ -47,7 +47,7 @@ export const getUsers = async (searchQuery) => {
 
 export const updateUser = async (userId, userBody, requesterId) => {
     const requester = await getUser(requesterId);
-    if (requester.admin || requester._id == userId) {
+    if (requester && (requester.admin || requester._id == userId)) {
         const updatedDate = new Date(Date.now());
         const updateBody = {
             ...userBody,
@@ -60,14 +60,18 @@ export const updateUser = async (userId, userBody, requesterId) => {
 
 export const deleteUser = async (userId, requesterId) => {
     const requester = await getUser(requesterId);
-    if (requester.admin || requester._id == userId) {
+    if (requester && (requester.admin || requester._id == userId)) {
         return await (User.findByIdAndDelete({ _id: userId }));
     }
     throw Error('Action not authorised');
 }
 
-export const deleteUsers = async () => {
-    return await (User.deleteMany());
+export const deleteUsers = async (requesterId) => {
+    const requester = await getUser(requesterId);
+    if (requester && requester.admin) {
+        return await (User.deleteMany());
+    }
+    throw Error('Action not authorised');
 }
 
 const encryptPassword = async (password) => {
